refactor(signup): entity 리팩토링 및 API 응답 타입 분리
