+++ apps/web/src/entities/signup/api/dto/signup-post.dto.ts
@@ -1,0 +1,14 @@
+import type { ApiResponse } from '@/src/shared/api';
+
+export interface SignupRequestDTO {
+  email: string;
+  nickname: string;
+  password: string;
+  imagePath?: string;
+}
+
+export interface SignupDataType {
+  userId: number;
+}
+
+export type SignupResponseDTO = ApiResponse<SignupDataType>;
