+++ apps/web/src/entities/signup/index.ts
@@ -1,0 +1,1 @@
+export * from './api';
